[nativert] Move GraphExecutorBase to PyTorch core (#156196)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/156196

Moves GraphExecutorBase class to PyTorch core.
GraphExecutorBase is a lightweight abstraction to execute a graph with  execution frames without actually owning the graph nor the weights. This is introduced to decouple the state management of the top level runtime from the kernel executions so that sub graphs from higher order ops can be supported.

Torch Native Runtime RFC: pytorch/rfcs#72

Test Plan:
This is a base class. So CI is sufficient

Rollback Plan:

Reviewed By: zhxchen17

Differential Revision: D76830436